+++ R/constants.R
@@ -1,0 +1,2 @@
+##Various constants used in this package
+sd_const = (2.4^2)/4