+++ NEWS.md
@@ -1,3 +1,9 @@
+# Version 0.5.0 (21st June 2017)
+
+* Added functions for inference: `abc`, `mcmc`, `spsa`.
+
+* Added `fx` function which provides a demo analysis of exchange rate data.
+
 # Version 0.4.0 (2nd Feb 2017)
 
 * Complete rewrite to make code more efficient and stable. This involves changes to the argument required by the functions.